requirements.txt
python-telegram-bot==21.6
anthropic>=0.34.0
python-pptx>=0.6.23
matplotlib>=3.9.0
Pillow>=10.4.0
pandas>=2.2.0
openpyxl>=3.1.5
pyyaml>=6.0.2
numpy>=1.26.0
python-dotenv>=1.0.0
pypdf>=4.0.0
python-docx>=1.1.0
